fixed bug: severe problems if project is not loacted inside a view e.g. in snapshot views with multi VOB structure (did not cause any problems in dynamic views, as in a dynamic view all directories are in a view)

+++ net.sourceforge.eclipseccase.ui/src/net/sourceforge/eclipseccase/ui/actions/SetConfigSpecAction.java
@@ -60,11 +60,13 @@ public class SetConfigSpecAction extends ClearCaseWorkspaceAction {
 
 						ClearCaseInterface cci = ClearCase.createInterface(ClearCase.INTERFACE_CLI);
 						String viewName = cci.getViewName(resource.getLocation().toOSString());
-						cci.setViewConfigSpec(viewName, f.getPath(), new ConsoleOperationListener(monitor));
-						f.delete();
+						if (viewName.length() > 0) {
+							cci.setViewConfigSpec(viewName, f.getPath(), new ConsoleOperationListener(monitor));
+							f.delete();
 
-						if (resource != null) {
-							resource.getProject().refreshLocal(IResource.DEPTH_INFINITE, null);
+							if (resource != null) {
+								resource.getProject().refreshLocal(IResource.DEPTH_INFINITE, null);
+							}
 						}
 					} catch (Exception e) {
 						ClearCaseConsole console = ClearCaseConsoleFactory.getClearCaseConsole();
